refactor: layout row
