+++ webpage/src/views/CelesteView.vue
@@ -2,8 +2,11 @@
   <ViewLayout>
     <!-- LEFT CONTAINER -->
     <template #left-col>
-      <img alt="celeste game" src="https://gaming-cdn.com/images/products/8003/orig/celeste-pc-mac-juego-steam-cover.jpg?v=1705489821" class="image-display" style="width: 90%">
-      <el-row>
+      <LayoutRow>
+        <img alt="celeste game" src="https://gaming-cdn.com/images/products/8003/orig/celeste-pc-mac-juego-steam-cover.jpg?v=1705489821" class="image-display" style="width: 90%">
+      </LayoutRow>
+
+      <LayoutRow>
         <el-col :span="16">
           <p>
             {{ t('celeste.description') }}
@@ -12,7 +15,7 @@
         <el-col :span="8">
           <img alt="madeline" src="https://media.spelunky.fyi/mods/logo/01G8PXMPWZ2K3CVW70DJ5HEMGA/1658626661282822.gif" class="image-display" style="width: 90%">
         </el-col>
-      </el-row>
+      </LayoutRow>
     </template>
 
     <!-- CENTER CONTAINER -->
@@ -22,12 +25,12 @@
 
     <!-- RIGHT CONTAINER -->
     <template #right-col>
-      <el-col :span="24">
+      <LayoutRow>
         <CelesteComponent />
-      </el-col>
-      <el-row>
+      </LayoutRow>
+      <LayoutRow>
         <VideoCardComponent :video="'https://www.youtube.com/embed/_iBkkwy0fcM'" :title="'Cool speedy thingie'" :article="'yay me go fast'" ></VideoCardComponent>
-      </el-row>
+      </LayoutRow>
     </template>
 
   </ViewLayout>
@@ -37,7 +40,8 @@
 import CelesteComponent from "@/components/CelesteComponent.vue";
 import { useTranslation } from "i18next-vue";
 import VideoCardComponent from "@/components/VideoCardComponent.vue";
-import ViewLayout from "@/components/ViewLayout.vue";
+import ViewLayout from "@/components/layout/ViewLayout.vue";
+import LayoutRow from "@/components/layout/LayoutRow.vue";
 
 const { t } = useTranslation();
 </script>
